# docs/swagger/swagger.yaml
basePath: /api
definitions:
  realworld-endpoints_internal_dto.ArticleDTO:
    properties:
      author:
        $ref: '#/definitions/realworld-endpoints_internal_dto.ProfileDTO'
      body:
        type: string
      createdAt:
        type: string
      description:
        type: string
      favorited:
        type: boolean
      favoritesCount:
        type: integer
      slug:
        type: string
      tagList:
        items:
          type: string
        type: array
      title:
        type: string
      updatedAt:
        type: string
    type: object
  realworld-endpoints_internal_dto.ArticleResponse:
    properties:
      article:
        $ref: '#/definitions/realworld-endpoints_internal_dto.ArticleDTO'
    type: object
  realworld-endpoints_internal_dto.ArticlesResponse:
    properties:
      articles:
        items:
          $ref: '#/definitions/realworld-endpoints_internal_dto.ArticleDTO'
        type: array
      articlesCount:
        type: integer
    type: object
...
  realworld-endpoints_internal_dto.ProfileDTO:
    properties:
      bio:
        type: string
      following:
        type: boolean
      image:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend API implementing the RealWorld spec built with Go, Echo Framework,
    GORM, and PostgreSQL.
  title: RealWorld Conduit API
  version: "1.0"
paths:
  /articles:
    get:
      description: List articles with optional filtering by tag, author, favorited
        user, and pagination
      parameters:
      - description: Filter by tag
        in: query
        name: tag
        type: string
      - description: Filter by author username
        in: query
        name: author
        type: string
      - description: Filter by favorited username
        in: query
        name: favorited
        type: string
      - description: Limit number of articles (default 20)
        in: query
        name: limit
        type: integer
      - description: Offset/skip number of articles (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/realworld-endpoints_internal_dto.ArticlesResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List articles
      tags:
      - Articles
  /articles/{slug}:
    get:
      description: Get a single article by its slug
      parameters:
      - description: Article slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/realworld-endpoints_internal_dto.ArticleResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get article by slug
      tags:
      - Articles
...
  /articles/{slug}/favorite:
    delete:
      description: Remove an article from the current user's favorites
      parameters:
      - description: Article slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/realworld-endpoints_internal_dto.ArticleResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Unfavorite an article
      tags:
      - Favorites
    post:
      description: Mark an article as favorite for the current user
      parameters:
      - description: Article slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/realworld-endpoints_internal_dto.ArticleResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Favorite an article
      tags:
      - Favorites
  /articles/feed:
    get:
      description: Get articles from users that the current user follows
      parameters:
      - description: Limit number of articles (default 20)
        in: query
        name: limit
        type: integer
      - description: Offset/skip number of articles (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/realworld-endpoints_internal_dto.ArticlesResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Feed articles
      tags:
      - Articles
...
securityDefinitions:
  ApiKeyAuth:
    description: 'JWT Token. Format: "Token {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"